I made the openMP for Jacobi

+++ jacobi.h
@@ -1,6 +1,6 @@
 #ifndef __JACOBI_H
 #define __JACOBI_H
 
-int jacobi(int N, double **array_in, double **array_out, int max_iter, double threshold, double **fmatrix);
+double jacobi(int N, double **array_in, double **array_out, int max_iter, double threshold, double **fmatrix);
 
 #endif
